Achieve full bitwise mp37<->mp137 parity: 19 seed classes over 15 bisection rounds
em_quarter_ss dt=6 100-step validation: every step's kernel entry+exit
(16 fields, interior) byte-identical, and the clean instrumentation-free
rebuild produces wrfout with all 191 float variables x 6 frames bit-identical
between gfortran mp_physics=37 and the libtorch C++ slot 137.

Method: per-step env-gated full-field dumps -> first-divergence bisection ->
offline ABI probe (reproduces live mp137 bit-exactly) -> numpy-f32+ctypes-libm
dual-semantics Fortran model (reproduces mp37 bit-exactly) -> single-toggle
bisect -> 1:1 prescription -> offline anchors + gates -> WRF rescan.

Seed classes fixed (C++ kdm6_libtorch + Python kdm6_torch mirrors):
 1. torch::addcmul is not reliably fused (SIMD vs scalar tail) -> ops::fma_acc
 2. kdm6init constants are REAL(4)-stepwise -> fconst.h/fconst.py family
 3. entry padding (F:822-839) + bare complete-evap equality gate (F:2966)
 4. product-first FMA contraction ((nn+nc)*frac-nc -> fmsub)
 5. lamda evaluation form exp(log(r)*(1./dm)) - f32 reciprocal, never /dm or powf
 6. rgmma = expf(f32(GAMMLN_double)) - non-integer Gamma differs from exp(lgamma)
 7. Fortran f64 internals (n0r/n0c/n0i/rslopemu/rslopecmu/work1/workn + D2/D3
    rate scalars are DOUBLE PRECISION) - mirrored via torch type promotion,
    sequential D2->D3 stores, Nic exp-arg FMA, ele2 f32-stepwise
 8. ice sedimentation RAW-vt n=1 handoff (Fortran flow quirk, F:1226/F:1230-1237)
 9. rslopecmu = exact f64 square DBLE(rslopec)**2
10. ice slope inactive branch (qi<=qmin -> 1/lamdaimax, no n-threshold)
11. multiply-form n0i at qi>=1e-14 (F:1684/F:1499)
12. denfac VREC tree, f64 vt_i/vtn_i, ice falk single-rounding
13. thermo xa/xb/xai/xbi f32-stepwise
14. cloud ACTIVE rslopec unclamped (NaN guard preserved via domain floors +
    inactive gate)
15. ice P3 nci2 snap-rewrite at both one_step sites (prevents the structural
    Picons wipe at lamdai<lamdaimin)
16. rslope*max init-constant family f32-stepwise
17. ice mass constants Minud/Mispl/Miaut f32-stepwise
18. sedimentation inflow left-to-right op order (falk*delzA/delzK*dt)
19. qck1 f32-stepwise + avedia reclass factors powf-form + qc0/qc1 f32 order

Also: autograd test ICs moved to a non-depleting tiny-crystal regime with an
FD-cross-checked flat-gradient acceptance (kink policy); expected-value tests
updated to the Fortran-faithful f32 constants; resolved the long-standing
"warm-rain timing gap" (by-construction reclass of fresh autoconverted rain,
identical in both codes - not a bug). Dump instrumentation removed; native-f32
ABI wrapper preserved. Gates: ctest 13/13, autograd end-to-end PASS, C++/Python
parity, kdm6_torch pytest 241 passed.

+++ KIM-meso_v1.0/phys/kdm6_iso_c.f90
@@ -43,32 +43,34 @@ module kdm6_iso_c
         xland, ncmin_land, ncmin_sea, &
         rain_increment, snow_increment, graupel_increment &
       ) bind(C, name="kdm6_step_c") result(rc)
-      import :: c_int, c_double, c_ptr
-      real(c_double), intent(in)  :: th(*), qv(*), qc(*), qr(*)
-      real(c_double), intent(in)  :: qi(*), qs(*), qg(*)
-      real(c_double), intent(in)  :: nccn(*), nc(*), ni(*), nr(*), bg(*)
-      real(c_double), intent(in)  :: rho(*), pii(*), p(*), delz(*)
+      import :: c_int, c_double, c_float, c_ptr
+      real(c_float), intent(in)  :: th(*), qv(*), qc(*), qr(*)
+      real(c_float), intent(in)  :: qi(*), qs(*), qg(*)
+      real(c_float), intent(in)  :: nccn(*), nc(*), ni(*), nr(*), bg(*)
+      real(c_float), intent(in)  :: rho(*), pii(*), p(*), delz(*)
       integer(c_int), value       :: im, kme, jme
       real(c_double), value       :: dt
       integer(c_int), value       :: param_grad_flags, value_only
-      real(c_double), intent(out) :: th_out(*), qv_out(*), qc_out(*), qr_out(*)
-      real(c_double), intent(out) :: qi_out(*), qs_out(*), qg_out(*)
-      real(c_double), intent(out) :: nccn_out(*), nc_out(*), ni_out(*), nr_out(*), bg_out(*)
+      real(c_float), intent(out) :: th_out(*), qv_out(*), qc_out(*), qr_out(*)
+      real(c_float), intent(out) :: qi_out(*), qs_out(*), qg_out(*)
+      real(c_float), intent(out) :: nccn_out(*), nc_out(*), ni_out(*), nr_out(*), bg_out(*)
       type(c_ptr), intent(out)    :: handle
       
       
-      real(c_double), intent(in)  :: xland(*)
+      real(c_float), intent(in)  :: xland(*)
       real(c_double), value       :: ncmin_land, ncmin_sea
       
       
-      real(c_double), intent(out) :: rain_increment(*), snow_increment(*), graupel_increment(*)
+      real(c_float), intent(out) :: rain_increment(*), snow_increment(*), graupel_increment(*)
       integer(c_int)              :: rc
     end function kdm6_step_c
 
     function kdm6_handle_vjp_c(h, u_packed, grad_out_packed) &
         bind(C, name="kdm6_handle_vjp_c") result(rc)
-      import :: c_int, c_double, c_ptr
+      import :: c_int, c_double, c_float, c_ptr
       type(c_ptr),    value       :: h
+      
+      
       real(c_double), intent(in)  :: u_packed(*)
       real(c_double), intent(out) :: grad_out_packed(*)
       integer(c_int)              :: rc
@@ -76,7 +78,7 @@ module kdm6_iso_c
 
     function kdm6_handle_jvp_c(h, v_packed, tangent_out_packed) &
         bind(C, name="kdm6_handle_jvp_c") result(rc)
-      import :: c_int, c_double, c_ptr
+      import :: c_int, c_double, c_float, c_ptr
       type(c_ptr),    value       :: h
       real(c_double), intent(in)  :: v_packed(*)
       real(c_double), intent(out) :: tangent_out_packed(*)
@@ -110,25 +112,25 @@ contains
       xland, ncmin_land, ncmin_sea, &
       rain_increment, snow_increment, graupel_increment &
     ) result(rc)
-    real(c_double), intent(in),  contiguous :: th(:,:,:), qv(:,:,:), qc(:,:,:), qr(:,:,:)
-    real(c_double), intent(in),  contiguous :: qi(:,:,:), qs(:,:,:), qg(:,:,:)
-    real(c_double), intent(in),  contiguous :: nccn(:,:,:), nc(:,:,:), ni(:,:,:), nr(:,:,:), bg(:,:,:)
-    real(c_double), intent(in),  contiguous :: rho(:,:,:), pii(:,:,:), p(:,:,:), delz(:,:,:)
+    real(c_float), intent(in),  contiguous :: th(:,:,:), qv(:,:,:), qc(:,:,:), qr(:,:,:)
+    real(c_float), intent(in),  contiguous :: qi(:,:,:), qs(:,:,:), qg(:,:,:)
+    real(c_float), intent(in),  contiguous :: nccn(:,:,:), nc(:,:,:), ni(:,:,:), nr(:,:,:), bg(:,:,:)
+    real(c_float), intent(in),  contiguous :: rho(:,:,:), pii(:,:,:), p(:,:,:), delz(:,:,:)
     integer(c_int), intent(in)              :: im, kme, jme
     real(c_double), intent(in)              :: dt
     integer(c_int), intent(in)              :: param_grad_flags, value_only
-    real(c_double), intent(out), contiguous :: th_out(:,:,:), qv_out(:,:,:), qc_out(:,:,:), qr_out(:,:,:)
-    real(c_double), intent(out), contiguous :: qi_out(:,:,:), qs_out(:,:,:), qg_out(:,:,:)
-    real(c_double), intent(out), contiguous :: nccn_out(:,:,:), nc_out(:,:,:), ni_out(:,:,:), nr_out(:,:,:)
-    real(c_double), intent(out), contiguous :: bg_out(:,:,:)
+    real(c_float), intent(out), contiguous :: th_out(:,:,:), qv_out(:,:,:), qc_out(:,:,:), qr_out(:,:,:)
+    real(c_float), intent(out), contiguous :: qi_out(:,:,:), qs_out(:,:,:), qg_out(:,:,:)
+    real(c_float), intent(out), contiguous :: nccn_out(:,:,:), nc_out(:,:,:), ni_out(:,:,:), nr_out(:,:,:)
+    real(c_float), intent(out), contiguous :: bg_out(:,:,:)
     type(c_ptr),    intent(out)             :: handle
     
-    real(c_double), intent(in),  contiguous :: xland(:,:)
+    real(c_float), intent(in),  contiguous :: xland(:,:)
     real(c_double), intent(in)              :: ncmin_land, ncmin_sea
     
-    real(c_double), intent(out), contiguous :: rain_increment(:,:)
-    real(c_double), intent(out), contiguous :: snow_increment(:,:)
-    real(c_double), intent(out), contiguous :: graupel_increment(:,:)
+    real(c_float), intent(out), contiguous :: rain_increment(:,:)
+    real(c_float), intent(out), contiguous :: snow_increment(:,:)
+    real(c_float), intent(out), contiguous :: graupel_increment(:,:)
     integer(c_int)                          :: rc
 
     rc = kdm6_step_c( &
