+++ kdm6_libtorch/bridge/kdm6_c_api.cpp
@@ -68,24 +68,24 @@ void ensure_libtorch_singlethread() {
 }  // anonymous namespace
 
 extern "C" int kdm6_step_c(
-    const double* th, const double* qv, const double* qc, const double* qr,
-    const double* qi, const double* qs, const double* qg,
-    const double* nccn, const double* nc, const double* ni, const double* nr,
-    const double* bg,
-    const double* rho, const double* pii, const double* p, const double* delz,
+    const float* th, const float* qv, const float* qc, const float* qr,
+    const float* qi, const float* qs, const float* qg,
+    const float* nccn, const float* nc, const float* ni, const float* nr,
+    const float* bg,
+    const float* rho, const float* pii, const float* p, const float* delz,
     int im, int kme, int jme, double dt,
     int param_grad_flags, int value_only,
-    double* th_out, double* qv_out, double* qc_out, double* qr_out,
-    double* qi_out, double* qs_out, double* qg_out,
-    double* nccn_out, double* nc_out, double* ni_out, double* nr_out,
-    double* bg_out,
+    float* th_out, float* qv_out, float* qc_out, float* qr_out,
+    float* qi_out, float* qs_out, float* qg_out,
+    float* nccn_out, float* nc_out, float* ni_out, float* nr_out,
+    float* bg_out,
     kdm6_handle_t** handle,
-    const double* xland,
+    const float* xland,
     double ncmin_land,
     double ncmin_sea,
-    double* rain_increment,
-    double* snow_increment,
-    double* graupel_increment
+    float* rain_increment,
+    float* snow_increment,
+    float* graupel_increment
 ) {
     if (im <= 0 || kme <= 0 || jme <= 0) return KDM6_ERR_INVALID_DIM;
     // Note: `xland` deliberately excluded from null check — it's optional.
@@ -110,13 +110,13 @@ extern "C" int kdm6_step_c(
         auto params = kdm6::make_parameters(param_grad_flags);
 
         // Optional xland → Tensor conversion (state.cpp layout convention):
-        // Fortran xland(im, jme) is column-major double*; C-order view is
+        // Fortran xland(im, jme) is column-major float*; C-order view is
         // (jme, im); permute({1, 0}) gives (im, jme); reshape to flat
         // (im*jme,) batch matches state batch index = i*jme + j.
         c10::optional<torch::Tensor> xland_t = c10::nullopt;
         if (xland != nullptr) {
-            auto opts = torch::TensorOptions().dtype(torch::kFloat64);
-            auto view2d = torch::from_blob(const_cast<double*>(xland),
+            auto opts = torch::TensorOptions().dtype(torch::kFloat32);
+            auto view2d = torch::from_blob(const_cast<float*>(xland),
                                            {jme, im}, opts)
                               .permute({1, 0})
                               .contiguous();
@@ -131,15 +131,15 @@ extern "C" int kdm6_step_c(
                                 nccn_out, nc_out, ni_out, nr_out, bg_out);
 
         // Copy-back sedimentation surface increments — (im*jme,) Tensor →
-        // Fortran column-major double*(im, jme). Reuses the same (im, jme) ↔
+        // Fortran column-major float*(im, jme). Reuses the same (im, jme) ↔
         // batch-flat layout convention as xland staging (state.cpp:67-72):
         //   batch index B = i*jme + j; Fortran memory = (i, j) column-major.
         // So tensor[B] at index i*jme+j maps to fortran[i + im*j].
-        auto copy_increment = [im, jme](const torch::Tensor& inc, double* dst) {
+        auto copy_increment = [im, jme](const torch::Tensor& inc, float* dst) {
             if (dst == nullptr) return;
-            auto inc_cpu = inc.contiguous().to(torch::kCPU, torch::kFloat64);
+            auto inc_cpu = inc.contiguous().to(torch::kCPU, torch::kFloat32);
             auto inc_2d = inc_cpu.reshape({im, jme}); // (im, jme) C-order
-            const double* src = inc_2d.data_ptr<double>();
+            const float* src = inc_2d.data_ptr<float>();
             // Transpose to Fortran (im, jme) column-major: dst[i + im*j] = src[i*jme + j]
             for (int i = 0; i < im; ++i) {
                 for (int j = 0; j < jme; ++j) {
