+++ kdm6_libtorch/bridge/kdm6_c_api.h
@@ -27,8 +27,8 @@ enum {
 /**
  * 한 step KDM6 호출. Fortran forward와 *동반 구동*되어 derivative 정보 산출.
  *
- * 입력 state/forcing은 Fortran-allocated double*(im, kme, jme).
- * 출력 state는 caller가 *미리 할당*한 double*(im, kme, jme)에 결과 복사.
+ * 입력 state/forcing은 Fortran-allocated float*(im, kme, jme).
+ * 출력 state는 caller가 *미리 할당*한 float*(im, kme, jme)에 결과 복사.
  *
  * @param param_grad_flags  Bitwise OR of:
  *                            1 = PEAUT, 2 = NCRK1, 4 = NCRK2, 8 = ECCBRK, 15 = ALL
@@ -40,43 +40,43 @@ enum {
  */
 int kdm6_step_c(
     /* in: 12 prognostic state */
-    const double* th, const double* qv, const double* qc, const double* qr,
-    const double* qi, const double* qs, const double* qg,
-    const double* nccn, const double* nc, const double* ni, const double* nr,
-    const double* bg,
+    const float* th, const float* qv, const float* qc, const float* qr,
+    const float* qi, const float* qs, const float* qg,
+    const float* nccn, const float* nc, const float* ni, const float* nr,
+    const float* bg,
     /* in: 4 forcing */
-    const double* rho, const double* pii, const double* p, const double* delz,
+    const float* rho, const float* pii, const float* p, const float* delz,
     /* in: dimensions, dt, flags */
     int im, int kme, int jme, double dt,
     int param_grad_flags, int value_only,
     /* out: 12 prognostic */
-    double* th_out, double* qv_out, double* qc_out, double* qr_out,
-    double* qi_out, double* qs_out, double* qg_out,
-    double* nccn_out, double* nc_out, double* ni_out, double* nr_out,
-    double* bg_out,
+    float* th_out, float* qv_out, float* qc_out, float* qr_out,
+    float* qi_out, float* qs_out, float* qg_out,
+    float* nccn_out, float* nc_out, float* ni_out, float* nr_out,
+    float* bg_out,
     /* out: opaque handle */
     kdm6_handle_t** handle,
     /* in (optional, appended): land/sea mask + per-regime ncmin scalars.
-     * `xland` is a double*(im, jme) — Fortran 2-D pattern matches WRF XLAND
+     * `xland` is a float*(im, jme) — Fortran 2-D pattern matches WRF XLAND
      * (>=1.5 → sea, else land). May be NULL → C++ falls back to the
      * pre-extension behavior (sea_mask = all true, scalar `constants::NCMIN`).
      * When non-NULL, `ncmin_land` / `ncmin_sea` are applied per-cell:
      *   ncmin_eff(i,j) = (xland(i,j) >= 1.5) ? ncmin_sea : ncmin_land
      * and injected into the C++ Phase Params' `ncmin_tensor` field.
      * See module_mp_kdm6ad.F (caller) and runtime.cpp (consumer). */
-    const double* xland,
+    const float* xland,
     double ncmin_land,
     double ncmin_sea,
     /* out (optional, appended): per-column sedimentation surface increments
-     * [mm] for the timestep `dt`. Each is a Fortran-allocated double*(im, jme)
+     * [mm] for the timestep `dt`. Each is a Fortran-allocated float*(im, jme)
      * (column-major) or NULL to discard. WRF wrapper uses these to accumulate
      * RAINNCV / SNOWNCV / GRAUPELNCV / SR / RAIN / SNOW / GRAUPEL.
      * If NULL, the runtime still computes sedimentation but the increments
      * are not copied out. See module_mp_kdm6ad.F:1304-1324 reference for the
      * Fortran-side accumulation pattern. */
-    double* rain_increment,
-    double* snow_increment,
-    double* graupel_increment
+    float* rain_increment,
+    float* snow_increment,
+    float* graupel_increment
 );
 
 /**
@@ -86,13 +86,18 @@ int kdm6_step_c(
  * `grad_out_packed`도 동일 layout, caller-allocated.
  *
  * @return KDM6_OK 또는 KDM6_ERR_VALUE_ONLY / HANDLE_CLOSED / NOT_IMPLEMENTED.
+ *
+ * NOTE: the derivative (VJP/JVP) packed arrays stay `double` ON PURPOSE — only the
+ * operational state/forcing ABI (kdm6_step_c) went native float32 to match Fortran
+ * mp37 (RWORDSIZE=4). 4D-Var adjoint/tangent precision must remain fp64 (the
+ * differentiable oracle's precision); float32 gradient buffers would degrade it.
  */
 int kdm6_handle_vjp_c(kdm6_handle_t* h,
                       const double* u_packed,
                       double* grad_out_packed);
 
 /**
- * JVP — J @ v. EnKF perturbation 용.
+ * JVP — J @ v. EnKF perturbation 용. (packed arrays fp64 — see VJP note above.)
  */
 int kdm6_handle_jvp_c(kdm6_handle_t* h,
                       const double* v_packed,
